requirements.txt
uvicorn[standard]==0.24.0
alembic==1.13.0
python-multipart==0.0.6
supabase==2.3.0
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
fastapi==0.116.1
idna==3.10
iniconfig==2.1.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pip==24.2
pluggy==1.6.0
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
ruff==0.13.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.43
starlette==0.47.3
structlog==25.4.0
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0